pyproject.toml
[tool.poetry]
name = "sdrdm_database"
version = "0.4.1"
description = "An sdRDM-based framework to utilize databases"
authors = ["Jan Range <30547301+JR-1991@users.noreply.github.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
ibis-framework = {extras = ["mysql"], version = "^7.0.0"}
gitpython = "^3.1.40"
validators = "^0.22.0"
strawberry-graphql = "^0.209.8"
joblib = "^1.3.2"
typer = "^0.9.0"
pydantic = "2.5.3"
sdrdm = "^0.2.2"

[tool.poetry.extras]
mysql = ["sqlalchemy", "pymysql", "sqlalchemy-views"]
postgres = ["psycopg2", "sqlalchemy", "sqlalchemy-views"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[tool.poetry.scripts]
sdrdm-db = "sdrdm_database.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = ["integration"]
